refactor(turbo_payment_bloc): change _currentAmount type from double to int for better precision feat(presetAmountSelector): add onAmountSelected callback to handle custom amount changes style(top_up_dialog): adjust layout for mobile and desktop using ScreenTypeLayout.builder
refactor(turbo_payment_bloc_test.dart): change selectedAmount type from double to int to improve consistency and avoid unnecessary type conversions

+++ lib/blocs/turbo_payment/turbo_payment_bloc.dart
@@ -32,10 +32,10 @@ class PaymentBloc extends Bloc<PaymentEvent, PaymentState> {
 
   FileSizeUnit _currentDataUnit = FileSizeUnit.gigabytes;
   String _currentCurrency = 'usd';
-  double _currentAmount = presetAmounts.first.toDouble();
+  int _currentAmount = presetAmounts.first;
 
   String get currentCurrency => _currentCurrency;
-  double get currentAmount => _currentAmount;
+  int get currentAmount => _currentAmount;
   FileSizeUnit get currentDataUnit => _currentDataUnit;
 
   late BigInt costOfOneGb;
@@ -96,7 +96,7 @@ class PaymentBloc extends Bloc<PaymentEvent, PaymentState> {
 
   _recomputePriceEstimate(
     Emitter emit, {
-    required double currentAmount,
+    required int currentAmount,
     required String currentCurrency,
   }) async {
     emit(PaymentLoading());
