+++ test/blocs/turbo_payment_bloc_test.dart
@@ -57,7 +57,7 @@ void main() {
       PaymentLoaded(
         balance: BigInt.from(oneAR * 2),
         estimatedStorageForBalance: '2.00',
-        selectedAmount: 25.0,
+        selectedAmount: 25,
         creditsForSelectedAmount: BigInt.from(oneAR),
         estimatedStorageForSelectedAmount: '1.00',
         currencyUnit: 'usd',
@@ -75,20 +75,19 @@ void main() {
         () => mockPaymentService.getBalance(wallet: wallet),
       ).thenAnswer((_) async => BigInt.from(oneAR * 2));
       when(
-        () =>
-            mockPaymentService.getPriceForFiat(currency: 'usd', amount: 150.0),
+        () => mockPaymentService.getPriceForFiat(currency: 'usd', amount: 150),
       ).thenAnswer((_) async => BigInt.from(oneAR * 3));
     },
     build: () {
       return paymentBloc;
     },
-    act: (bloc) => bloc.add(FiatAmountSelected(150.0)),
+    act: (bloc) => bloc.add(FiatAmountSelected(150)),
     expect: () => [
       PaymentLoading(),
       PaymentLoaded(
         balance: BigInt.from(oneAR * 2),
         estimatedStorageForBalance: '2.00',
-        selectedAmount: 150.0,
+        selectedAmount: 150,
         creditsForSelectedAmount: BigInt.from(oneAR * 3),
         estimatedStorageForSelectedAmount: '3.00',
         currencyUnit: 'usd',
@@ -105,7 +104,7 @@ void main() {
           .thenAnswer((_) async => BigInt.from(oneAR));
       when(() => mockPaymentService.getPriceForFiat(
             currency: 'cad',
-            amount: presetAmounts.first.toDouble(),
+            amount: presetAmounts.first,
           )).thenAnswer((_) async => BigInt.from(oneAR * 0.5));
       when(() =>
               mockPaymentService.getPriceForBytes(byteSize: oneGigbyteInBytes))
@@ -120,7 +119,7 @@ void main() {
       PaymentLoaded(
         balance: BigInt.from(oneAR),
         estimatedStorageForBalance: '1.00',
-        selectedAmount: presetAmounts.first.toDouble(),
+        selectedAmount: presetAmounts.first,
         creditsForSelectedAmount: BigInt.from(oneAR * 0.5),
         estimatedStorageForSelectedAmount: '0.50',
         currencyUnit: 'cad',
@@ -137,7 +136,7 @@ void main() {
       when(() => mockPaymentService.getBalance(wallet: wallet))
           .thenAnswer((_) async => BigInt.from(oneAR));
       when(() =>
-              mockPaymentService.getPriceForFiat(currency: 'usd', amount: 25.0))
+              mockPaymentService.getPriceForFiat(currency: 'usd', amount: 25))
           .thenAnswer((_) async => BigInt.from(oneAR));
       when(() =>
               mockPaymentService.getPriceForBytes(byteSize: oneGigbyteInBytes))
@@ -153,7 +152,7 @@ void main() {
         PaymentLoaded(
           balance: BigInt.from(oneAR),
           estimatedStorageForBalance: '1024.00',
-          selectedAmount: 25.0,
+          selectedAmount: 25,
           creditsForSelectedAmount: BigInt.from(oneAR),
           estimatedStorageForSelectedAmount: '1024.00',
           currencyUnit: 'usd',
